+++ src/routes/user.ts
@@ -1,7 +1,10 @@
 import { Router } from 'express'
 
-import { deleteUser, getUser, getUsers, putUser } from '@controllers/user'
 import { createUser } from '@controllers/User/createUser'
+import { getUser } from '@controllers/User/getUser'
+import { getUsers } from '@controllers/User/getUsers'
+import { deleteUser } from '@controllers/User/deleteUser'
+import { putUser } from '@controllers/User/putUser'
 
 const userRouter = Router()
